test: move lint_data tests alongside the module
